requirements.txt
fastapi
httpx
mediapipe
numpy
pillow
pytest
uvicorn